big cleanup creating branch minimise
