+++ src/NamedPlus.jl
@@ -3,23 +3,22 @@ module NamedPlus
 using NamedDims
 export NamedDims, NamedDimsArray, dim, unname
 
-using AxisRanges
-export wrapdims, RangeArray, ranges
-
-include("wrap.jl")
-
-export NamedUnion
+# using AxisRanges # perhaps they can co-operate for now
+# export wrapdims, RangeArray, ranges
 
 include("recursion.jl")
-
 export getnames, nameless
-
-include("view.jl")
 
 include("macro.jl")
 
-include("maths.jl")
+include("permute.jl")
 
-include("plots.jl")
+include("create.jl")
+
+include("reshape.jl")
+
+include("rename.jl")
+
+# include("plots.jl")
 
 end # module
